+++ umac/regulatory/dispatcher/inc/wlan_reg_services_api.h
@@ -1618,7 +1618,7 @@ wlan_reg_chan_has_dfs_attribute_for_freq(struct wlan_objmgr_pdev *pdev,
 QDF_STATUS
 wlan_reg_get_channel_list_with_power_for_freq(struct wlan_objmgr_pdev *pdev,
 					      struct channel_power *ch_list,
-					      uint8_t *num_chan);
+					      uint16_t *num_chan);
 
 #ifdef CONFIG_REG_6G_PWRMODE
 /**
