+++ umac/scan/dispatcher/inc/wlan_scan_public_structs.h
@@ -1021,7 +1021,7 @@ struct chan_list_scan_info {
  * @chan: channel parameters used for this scan
  */
 struct chan_list {
-	uint8_t num_chan;
+	uint16_t num_chan;
 	struct chan_info chan[NUM_CHANNELS];
 };
